Introducing ContainerAwareTestCase as superclass of BundleTestCase and ExtensionTestCase
