+++ Test/BundleTestCase.php
@@ -5,39 +5,12 @@
 
 namespace IC\Bundle\Base\TestBundle\Test;
 
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use IC\Bundle\Base\TestBundle\Test\TestCase;
-
 /**
  * Bundle Unit test case
  *
  * @author Guilherme Blanco <guilhermeblanco@hotmail.com>
+ * @author Ryan Albon <ryanalbon@gmail.com>
  */
-abstract class BundleTestCase extends TestCase
+abstract class BundleTestCase extends ContainerAwareTestCase
 {
-    /**
-     * @var \Symfony\Component\DependencyInjection\ContainerBuilder
-     */
-    protected $container;
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
-    {
-        parent::setUp();
-
-        $this->container = new ContainerBuilder();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function tearDown()
-    {
-        unset($this->container);
-
-        parent::tearDown();
-    }
 }
